add simple details page

+++ tbdd_pdm_core/api/details.py
@@ -8,12 +8,28 @@ import sqlalchemy.exc
 def create_detail(session, **kwargs):
     try:
         try:
-            session.add(models.Detail(
+            detail = models.Detail(
                 **kwargs
-            ))
+            )
+            session.add(detail)
             session.commit()
+            return detail.get_dict_fields()
         except sqlalchemy.exc.IntegrityError:
             raise DetailAlreadyExistsException(kwargs['code'])
+    finally:
+        session.rollback()
+
+
+def update_detail(session, id=None, name=None, description=None, is_standard=False, **kwargs):
+    try:
+        detail = session.query(models.Detail).filter(models.Detail.code == id).first()
+        if detail is None:
+            raise DetailNotFoundException(id)
+        detail.name = name
+        detail.description = description
+        detail.is_standard = is_standard
+        session.commit()
+        return detail.get_dict_fields()
     finally:
         session.rollback()
 
@@ -21,6 +37,8 @@ def create_detail(session, **kwargs):
 def delete_detail(session, detail_id=None, **kwargs):
     try:
         detail = session.query(models.Detail).filter(models.Detail.id == detail_id).first()
+        if detail is None:
+            raise DetailNotFoundException(id)
         session.delete(detail)
         session.commit()
     finally:
@@ -72,3 +90,28 @@ def get_assembly_tree(session, parent_id=None):
     finally:
         session.rollback()
 
+
+def get_list(session, results_per_page=20, page=1, filter=None):
+    result = {}
+    try:
+        total_count = session.query(models.Detail).count()
+        total_pages = int(total_count/results_per_page)+1
+        result['total_count'] = total_count
+        result['total_pages'] = total_pages
+        result['data'] = []
+        for detail in session.query(
+                models.Detail
+        ).offset(
+            (page-1)*results_per_page
+        ).limit(results_per_page):
+            result['data'].append(detail.get_dict_fields())
+        return result
+    finally:
+        session.rollback()
+
+
+def get_detail_by_id(session, detail_id=None):
+    detail = session.query(models.Detail).filter(models.Detail.id == detail_id).first()
+    if detail is None:
+        raise DetailNotFoundException
+    return detail.get_dict_fields()
